Replace browser test for action plugins to kernel test.
